Fix hiding a lifetime that's named elsewhere is confusing lints
The lifetime of CRCConfiguration is tied to the name of the
configuration.

Right now, none of the code uses the CRC name.  But there's a stash
that has Debug traits that uses it.

The other option was a stack-based string crate to remove the need for
lifetimes, but I wanted to keep the code simple for now.

Also fixed a empty line after doc comment lint.