Rename CRCEndianness to BitOrder and specify first to bit order
This is an API breaking change but it is necessary because I was using
terminology wrong and it was confusing.

Previously, I was using the terms MSB and LSB alone to describe CRC
configurations.  Now I am using MSBFirst and LSBFirst.  Also, instead
of CRCEndianness it is now BitOrder.  The CRC prefix is handled by the
module namespace and BitOrder more accurately reflects what the
variants are.